Add some python pika examples
This adds a python pika example and how to marsall things in and
out. The consumer and producer are supposed to run in parallel,
the consumer is started first then the producer can be run any
number of times after that.

Even though this is a basic example of using pika with RabbitMQ,
there is still an open question about verifying the spec. Since
pika is wrapping both the input and output of the code, I have no
way to verify the data going in/out of the protocol matches the
cloud event spec.

Signed-off-by: David Brown <dmlb2000@gmail.com>